[dns] rename to 'dns_types.hpp'
This commit renames the DNS helper module from `dns_header` to
`dns_types`.
