+++ Android.bp
@@ -300,7 +300,7 @@ cc_library_static {
         "src/core/net/dhcp6_client.cpp",
         "src/core/net/dhcp6_server.cpp",
         "src/core/net/dns_client.cpp",
-        "src/core/net/dns_headers.cpp",
+        "src/core/net/dns_types.cpp",
         "src/core/net/dnssd_server.cpp",
         "src/core/net/icmp6.cpp",
         "src/core/net/ip6.cpp",
