+++ src/core/BUILD.gn
@@ -472,8 +472,8 @@ openthread_core_files = [
   "net/dhcp6_server.hpp",
   "net/dns_client.cpp",
   "net/dns_client.hpp",
-  "net/dns_headers.cpp",
-  "net/dns_headers.hpp",
+  "net/dns_types.cpp",
+  "net/dns_types.hpp",
   "net/dnssd_server.cpp",
   "net/dnssd_server.hpp",
   "net/icmp6.cpp",
